Делаю MovieListScreen

+++ app/src/main/java/com/example/kinopoiskapi/data/model/Rating.kt
@@ -2,7 +2,7 @@ package com.example.kinopoiskapi.data.model
 
 data class Rating(
     val await: Double,
-    val filmCritics: Int,
+    val filmCritics: Double,
     val imdb: Double,
     val kp: Double,
     val russianFilmCritics: Double,
